Revert "move TextFieldBase to the internal namespace"
This reverts commit 1339a610d2782bec2bad6b8efcde70802eb3b02d.
